CR#26982 Dynamic partition
